requirements.txt
fastapi==0.86.0
pandas==1.4.3
pydantic==1.9.0
PyJWT==2.6.0
starlette==0.17.1
uvicorn==0.14.0
yfinance==0.1.74
